rmin: really compute floats...
* vcsn/weightset/rmin.hh: Er... s/int/double/.

+++ vcsn/weightset/rmin.hh
@@ -12,10 +12,10 @@ namespace vcsn
   namespace detail
   {
     class rmin_impl
-      : public min_plus_impl<int>
+      : public min_plus_impl<double>
     {
     public:
-      using super_t = min_plus_impl<int>;
+      using super_t = min_plus_impl<double>;
       using self_t = rmin;
       using value_t = typename super_t::value_t;
